backend/requirements.txt
fastapi
uvicorn[standard]
pymongo>=4.3.3
requests
beautifulsoup4
python-dotenv
google-generativeai
selenium
